+++ packages/bridge-status-controller/src/bridge-status-controller.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-conditional-in-test */
 /* eslint-disable jest/no-restricted-matchers */
 import type { AccountsControllerActions } from '@metamask/accounts-controller';
 import { Messenger } from '@metamask/base-controller';
@@ -359,12 +358,14 @@ const MockTxHistory = {
   }),
   getPending: ({
     txMetaId = 'bridgeTxMetaId1',
+    batchId = undefined,
     srcTxHash = '0xsrcTxHash1',
     account = '0xaccount1',
     srcChainId = 42161,
     destChainId = 10,
   } = {}): Record<string, BridgeHistoryItem> => ({
     [txMetaId]: {
+      batchId,
       txMetaId,
       quote: getMockQuote({ srcChainId, destChainId }),
       startTime: 1729964825189,
@@ -457,6 +458,7 @@ const MockTxHistory = {
   }),
   getComplete: ({
     txMetaId = 'bridgeTxMetaId1',
+    batchId = undefined,
     srcTxHash = '0xsrcTxHash1',
     account = '0xaccount1',
     srcChainId = 42161,
@@ -464,6 +466,7 @@ const MockTxHistory = {
   } = {}): Record<string, BridgeHistoryItem> => ({
     [txMetaId]: {
       txMetaId,
+      batchId,
       quote: getMockQuote({ srcChainId, destChainId }),
       startTime: 1729964825189,
       completionTime: 1736277625746,
@@ -552,9 +555,12 @@ const executePollingWithPendingStatus = async () => {
     messenger: getMessengerMock(),
     clientId: BridgeClientId.EXTENSION,
     fetchFn: jest.fn(),
-    addTransactionFn: jest.fn(),
     estimateGasFeeFn: jest.fn(),
+    addTransactionBatchFn: jest.fn(),
     config: {},
+    state: {
+      txHistory: {},
+    },
   });
   const startPollingSpy = jest.spyOn(bridgeStatusController, 'startPolling');
 
@@ -588,8 +594,8 @@ const mockSelectedAccount = {
   },
 };
 
-const addTransactionFn = jest.fn();
 const estimateGasFeeFn = jest.fn();
+const addTransactionBatchFn = jest.fn();
 
 const getController = (call: jest.Mock, traceFn?: jest.Mock) => {
   const controller = new BridgeStatusController({
@@ -602,8 +608,8 @@ const getController = (call: jest.Mock, traceFn?: jest.Mock) => {
     } as never,
     clientId: BridgeClientId.EXTENSION,
     fetchFn: mockFetchFn,
-    addTransactionFn,
     estimateGasFeeFn,
+    addTransactionBatchFn,
     traceFn,
   });
 
@@ -627,8 +633,8 @@ describe('BridgeStatusController', () => {
         messenger: getMessengerMock(),
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       expect(bridgeStatusController.state).toStrictEqual(EMPTY_INIT_STATE);
       expect(mockMessengerSubscribe.mock.calls).toMatchSnapshot();
@@ -639,8 +645,9 @@ describe('BridgeStatusController', () => {
         messenger: getMessengerMock(),
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
         state: {
           txHistory: MockTxHistory.getPending(),
         },
@@ -675,8 +682,8 @@ describe('BridgeStatusController', () => {
         },
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       jest.advanceTimersByTime(10000);
       await flushPromises();
@@ -697,8 +704,12 @@ describe('BridgeStatusController', () => {
         messenger: getMessengerMock(),
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
+        state: {
+          txHistory: MockTxHistory.getPending(),
+        },
       });
 
       // Execution
@@ -734,8 +745,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       const fetchBridgeTxStatusSpy = jest.spyOn(
         bridgeStatusUtils,
@@ -814,8 +826,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
 
       // Start polling with args that have no srcTxHash
@@ -852,8 +865,8 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
 
       const fetchBridgeTxStatusSpy = jest
@@ -892,8 +905,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
 
       // Execution
@@ -954,8 +968,8 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
 
       // Start polling with no srcTxHash
@@ -1041,8 +1055,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       const fetchBridgeTxStatusSpy = jest
         .spyOn(bridgeStatusUtils, 'fetchBridgeTxStatus')
@@ -1127,8 +1142,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       const fetchBridgeTxStatusSpy = jest
         .spyOn(bridgeStatusUtils, 'fetchBridgeTxStatus')
@@ -1227,8 +1243,9 @@ describe('BridgeStatusController', () => {
         messenger: messengerMock,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
+
         estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
       });
       const fetchBridgeTxStatusSpy = jest
         .spyOn(bridgeStatusUtils, 'fetchBridgeTxStatus')
@@ -1461,7 +1478,7 @@ describe('BridgeStatusController', () => {
       expect(
         startPollingForBridgeTxStatusSpy.mock.lastCall[0],
       ).toMatchSnapshot();
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(controller.state.txHistory[result.id as never]).toMatchSnapshot();
     });
 
     it('should throw error when snap ID is missing', async () => {
@@ -1699,7 +1716,7 @@ describe('BridgeStatusController', () => {
 
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
       expect(result).toMatchSnapshot();
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(controller.state.txHistory[result.id as never]).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
     });
 
@@ -1865,12 +1882,33 @@ describe('BridgeStatusController', () => {
         gasFeeEstimates: { estimatedBaseFee: '0x1234' },
       });
       estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: mockApprovalTxMeta,
-        result: Promise.resolve('0xapprovalTxHash'),
-      });
       mockMessengerCall.mockReturnValueOnce({
-        transactions: [mockApprovalTxMeta],
+        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
+      });
+      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xapprovalTxHash',
+      });
+      mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
+    };
+
+    const setup2ApprovalMocks = () => {
+      mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
+      mockMessengerCall.mockReturnValueOnce('arbitrum-client-id');
+      mockMessengerCall.mockReturnValueOnce({
+        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
+      });
+      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
+      mockMessengerCall.mockReturnValueOnce({
+        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
+      });
+      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
+      mockMessengerCall.mockReturnValueOnce({
+        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
+      });
+      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xapprovalTxHash',
       });
     };
 
@@ -1881,12 +1919,8 @@ describe('BridgeStatusController', () => {
         gasFeeEstimates: { estimatedBaseFee: '0x1234' },
       });
       estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: mockEvmTxMeta,
-        result: Promise.resolve('0xevmTxHash'),
-      });
-      mockMessengerCall.mockReturnValueOnce({
-        transactions: [mockEvmTxMeta],
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xevmTxHash',
       });
 
       mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
@@ -1894,7 +1928,6 @@ describe('BridgeStatusController', () => {
 
     it('should successfully submit an EVM bridge transaction with approval', async () => {
       setupApprovalMocks();
-      setupBridgeMocks();
 
       const { controller, startPollingForBridgeTxStatusSpy } =
         getController(mockMessengerCall);
@@ -1903,8 +1936,10 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
 
@@ -1937,9 +1972,11 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
       expect(estimateGasFeeFn.mock.calls).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
 
@@ -1954,9 +1991,11 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
       expect(estimateGasFeeFn.mock.calls).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
 
@@ -1970,26 +2009,18 @@ describe('BridgeStatusController', () => {
       await expect(
         controller.submitTx(quoteWithoutApproval, false),
       ).rejects.toThrow(
-        'Failed to submit cross-chain swap transaction: unknown account in trade data',
+        'Failed to submit cross-chain swap batch transaction: unknown account in trade data',
       );
       controller.stopAllPolling();
 
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(addTransactionFn).not.toHaveBeenCalled();
+      expect(addTransactionBatchFn).not.toHaveBeenCalled();
     });
 
     it('should reset USDT allowance', async () => {
       mockIsEthUsdt.mockReturnValueOnce(true);
 
-      // USDT approval reset
-      mockMessengerCall.mockReturnValueOnce('1');
-      setupApprovalMocks();
-
-      // Approval tx
-      setupApprovalMocks();
-
-      // Bridge transaction
-      setupBridgeMocks();
+      setup2ApprovalMocks();
 
       const { controller, startPollingForBridgeTxStatusSpy } =
         getController(mockMessengerCall);
@@ -1998,98 +2029,12 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
       expect(estimateGasFeeFn.mock.calls).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
-    });
-
-    it('should throw an error if approval tx fails', async () => {
-      mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
-      mockMessengerCall.mockReturnValueOnce('arbitrum-client-id');
-      mockMessengerCall.mockReturnValueOnce({
-        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
-      });
-      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockRejectedValueOnce(new Error('Approval tx failed'));
-
-      const { controller, startPollingForBridgeTxStatusSpy } =
-        getController(mockMessengerCall);
-
-      await expect(
-        controller.submitTx(mockEvmQuoteResponse, false),
-      ).rejects.toThrow('Approval tx failed');
-
-      expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
-      expect(mockMessengerCall.mock.calls).toMatchSnapshot();
-    });
-
-    it('should throw an error if approval tx meta does not exist', async () => {
-      mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
-      mockMessengerCall.mockReturnValueOnce('arbitrum-client-id');
-      mockMessengerCall.mockReturnValueOnce({
-        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
-      });
-      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: undefined,
-        result: new Promise((resolve) => resolve('0xevmTxHash')),
-      });
-      mockMessengerCall.mockReturnValueOnce({
-        transactions: [],
-      });
-
-      setupBridgeMocks();
-      const { controller, startPollingForBridgeTxStatusSpy } =
-        getController(mockMessengerCall);
-
-      await expect(
-        controller.submitTx(mockEvmQuoteResponse, false),
-      ).rejects.toThrow(
-        'Failed to submit cross-chain swap tx: txMeta for txHash was not found',
-      );
-
-      expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
-      expect(mockMessengerCall.mock.calls).toMatchSnapshot();
-    });
-
-    it('should delay after submitting linea approval', async () => {
-      const handleLineaDelaySpy = jest
-        .spyOn(transactionUtils, 'handleLineaDelay')
-        .mockResolvedValueOnce();
-      const mockTraceFn = jest
-        .fn()
-        .mockImplementation((_p, callback) => callback());
-
-      setupApprovalMocks();
-      setupBridgeMocks();
-
-      const { controller, startPollingForBridgeTxStatusSpy } = getController(
-        mockMessengerCall,
-        mockTraceFn,
-      );
-
-      const lineaQuoteResponse = {
-        ...mockEvmQuoteResponse,
-        quote: { ...mockEvmQuoteResponse.quote, srcChainId: 59144 },
-        trade: {
-          ...(mockEvmQuoteResponse.trade as TxData),
-          gasLimit: undefined,
-        } as never,
-      };
-
-      const result = await controller.submitTx(lineaQuoteResponse, false);
-      controller.stopAllPolling();
-
-      expect(mockTraceFn).toHaveBeenCalledTimes(2);
-      expect(handleLineaDelaySpy).toHaveBeenCalledTimes(1);
-      expect(result).toMatchSnapshot();
-      expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
-      expect(mockMessengerCall.mock.calls).toMatchSnapshot();
-      expect(mockTraceFn.mock.calls).toMatchSnapshot();
     });
   });
 
@@ -2191,12 +2136,12 @@ describe('BridgeStatusController', () => {
         gasFeeEstimates: { estimatedBaseFee: '0x1234' },
       });
       estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: mockApprovalTxMeta,
-        result: Promise.resolve('0xapprovalTxHash'),
-      });
       mockMessengerCall.mockReturnValueOnce({
-        transactions: [mockApprovalTxMeta],
+        gasFeeEstimates: { estimatedBaseFee: '0x1234' },
+      });
+      estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xapprovalTxHash',
       });
     };
 
@@ -2207,12 +2152,8 @@ describe('BridgeStatusController', () => {
         gasFeeEstimates: { estimatedBaseFee: '0x1234' },
       });
       estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: mockEvmTxMeta,
-        result: Promise.resolve('0xevmTxHash'),
-      });
-      mockMessengerCall.mockReturnValueOnce({
-        transactions: [mockEvmTxMeta],
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xevmTxHash',
       });
 
       mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
@@ -2220,7 +2161,6 @@ describe('BridgeStatusController', () => {
 
     it('should successfully submit an EVM swap transaction with approval', async () => {
       setupApprovalMocks();
-      setupBridgeMocks();
 
       const { controller, startPollingForBridgeTxStatusSpy } =
         getController(mockMessengerCall);
@@ -2229,8 +2169,10 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
 
@@ -2263,9 +2205,11 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
       expect(estimateGasFeeFn.mock.calls).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
 
@@ -2276,9 +2220,8 @@ describe('BridgeStatusController', () => {
         gasFeeEstimates: { estimatedBaseFee: '0x1234' },
       });
       estimateGasFeeFn.mockResolvedValueOnce(mockEstimateGasFeeResult);
-      addTransactionFn.mockResolvedValueOnce({
-        transactionMeta: mockEvmTxMeta,
-        result: Promise.resolve('0xevmTxHash'),
+      addTransactionBatchFn.mockResolvedValueOnce({
+        batchId: '0xevmTxHash',
       });
       // mockMessengerCall.mockReturnValueOnce(mockSelectedAccount);
 
@@ -2290,9 +2233,11 @@ describe('BridgeStatusController', () => {
 
       expect(result).toMatchSnapshot();
       expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
-      expect(controller.state.txHistory[result.id]).toMatchSnapshot();
+      expect(
+        controller.state.txHistory[result.batchId as never],
+      ).toMatchSnapshot();
       expect(estimateGasFeeFn.mock.calls).toMatchSnapshot();
-      expect(addTransactionFn.mock.calls).toMatchSnapshot();
+      expect(addTransactionBatchFn.mock.calls).toMatchSnapshot();
       expect(mockMessengerCall.mock.calls).toMatchSnapshot();
     });
   });
@@ -2358,9 +2303,9 @@ describe('BridgeStatusController', () => {
         messenger: mockBridgeStatusMessenger,
         clientId: BridgeClientId.EXTENSION,
         fetchFn: jest.fn(),
-        addTransactionFn: jest.fn(),
-        estimateGasFeeFn: jest.fn(),
 
+        estimateGasFeeFn: jest.fn(),
+        addTransactionBatchFn: jest.fn(),
         state: {
           txHistory: {
             ...MockTxHistory.getPending(),